Web site map
Add a list of recently modified pages as a second tab on the site map
page.  I have not linked this page into the site navigation yet
because it seems to run rather slowly, so I need to fix that first,
which I think I can do by using caching.  I also filter out
robots.txt, which is not intended for user access.