Correct logb with on rounded
A call such as logb(3,2) does not evaluate to a numerical value with
on rounded.  Revise simplogb to call simplogbn only when dmode!* is
nil, which follows more closely the code of simplog.  This means that,
with on rounded, the simplogb code falls through to call simpiden,
which allows a numerical result to be returned.