Web REDUCE: add REDUCE and File menus
Add a REDUCE menu to control REDUCE and print a session log.  Replace
the <div> used previously for the I/O display with an <iframe>, which
is the only way I can find to print it including typeset maths, which
doesn't seem to copy using normal DOM methods.  Add a File menu that
currently only works on Chromium-based browsers as a first attempt at
providing file I/O, but input is not yet satisfactory.  Add some
experimental code that is currently hidden as an attempt to interleave
input and output, but this is currently very unreliable, probably
because I'm not handling the asynchronous nature of JavaScript
correctly.  I plan to upload this version to the REDUCE web site,
since at least the print facility appears to work and may be useful.