Web REDUCE: Makefile
Revise csl/new-embedded/for-emscripten/Makefile, primarily to be more
succinct: introduce variables to hold lists of object files to avoid
repeating identical or very similar lists; use multiple targets to
compress the rules for making object files. The new version should be
functionally identical to the old version but it has just over half as
many lines as the old version, and I think it's now easier to see the
essential structure. (I particularly want to understand the difference
between the node and browser builds because user queries work in the
node build but not in the browser build.)