Web REDUCE: delimiter matching
Highlight matching delimiter pairs in the input editor by underlining
the matching delimiters and their content when the text cursor is
immediately before an opening delimiter or immediately after a closing
delimiter.  Delimiter pairs matched are (...), {...}, <<...>> and
begin...end.  Matching is currently crude and just selects the nearest
appropriate delimiter, taking no account of nesting, strings,
comments, etc.  Delimiter matching can be turned off using the View
menu.  This is currently more a proof of concept than a seriously
useful facility.

To facilitate input text decoration I now use an editable <div> for
the input editor instead of a <textarea>.  The input editor code is
now in a JavaScript module and I use the extension .mjs for JavaScript
module files.  This requires an update to the .htaccess file to
declare an appropriate media type for .mjs files, which is still not
the default for the SourceForge web server.