Web REDUCE: fix a typo
Fix a typo in the iframe HTML caused when I moved it, because HTML and
JavaScript string syntax are slightly different, so I needed to double
the double backslashes!  Add more error handling for the web worker,
but it doesn't catch the error initialising REDUCE on mobile devices.
I think that's because the error is in a different thread, so it would
need to be caught within the worker.  That would be awkward to arrange
as the code is currently structured because the JavaScript wasm
wrapper is generated automatically by emscripten.  Delete the error
style definition from index.php that is now in the iframe code.