Web Site: SF badge and manual-lookup
Add the SourceForge "Open Source Excellence" badge to the footer.
This is the badge that also appears on our SourceForge project page.
Revise the manual-lookup facility to anchor the search string at the
start of a heading, because otherwise I can't disambiguate links to
some packages: in particular, "Matrix Calculations" appear in two
titles, but only at the start of one of them.