A quick fix for the bug reported in
 https://sourceforge.net/p/reduce-algebra/bugs/147/

After loading defint. the call to

  solve(x^(1/3)=i,x);

triggers a zero divisor error.

The cause is a rule for i^n, part of the ruleset ipower used for contour integration in
packages/defint/defintx.red.  In the groebner package which is used for solving polynomial
equations the normal rule i^2 => -1 is temporarily disabled, but the more complex one isn't.
Consequently, i^2 is simplified to 0 causing the error.

As a simple fix the ipower ruleset is now in effect only during contour integration, not globally
after loading the defint package. If someone were to add a rule for i^n by hand, the error would
show up again. 

A complete solution is certainly possible but needs careful study of the code in the dipoly package.