CSL: some more adjustments to help "--with-arith". One thing is that ariuth.h is patched so that that version will not transition to use of Karatsuba, since I have a bug there ar present. Also at present long (i.e. 128-bit) floating point reading and/or printing is all messed up. But those are both bugs I can look into later. But I am down to only about half a dozen of the main Reduce test files disagreeing with the reference output so I tend to feel fairly cheery!
General: At present the character "_" is a bit odd in the Reduce syntax in
that WITHING a token it is treated rather as a letter. So for instance the
sequnce multiple_word_symbol is treated as a single symbol. However at the
start of a token "_" is treated as punctuation, so if you want a symbol whose
name start with if you need to write something like !_underscored_. See that
the "_" at the end did not need a "!" but the one at the start does. OK that
is something that has been the state for AGES so radical change would not
be safe. However I would like to change things so that symbols can start
with a a double underscore without needing escape chars. This is in part
motivated by the C/C++ tradition that tokens starting with 2 underscores are
reserved for system use. If that was done one would be able to write eg
__file__ and __line__ just like that as well as __user_symbol. But things
like !_Single_underscore_at_start would still need the initial escape.
I would hope that this would be so close to upwards compatible and so few
people write two underscores in a row at present that it would not hurt
anybody. Objections please if you wish to object!