efjacinv.red:  Substantial improvements to the numerical evaluation routines of the inverse Jacobi functions. These should now return a value for all complex values of the argument and modulus (provided of course such a value exists). The value returned is now the principal value.
Work still needed to implement similar improvements in the numerical
evaluation of the incomplete elliptic integrals of the 1st and 2nd kinds
and Jacobi's Zeta function.

Minor alterations, corrections & deletions from other modules of the package.

ellipfn.tex & manual.pdf: A few paragraphs added to document the
improvements in the numerical evaluation of the Jacobi inverse functions.
Two new tables added to document the properties of the Jacobi functions
and their inverses.